Minor renaming: function `lulu_to_lstring()` -> `lulu_to_string`, macro `lulu_to_string` -> `lulu_to_cstring`
- Replace instances of `byte` with `u8` or `u32`, depending on context

+++ src/string.c
@@ -45,7 +45,7 @@ ostring_hash(const char *data, isize len)
 {
     u32 hash = FNV1A_OFFSET_32;
     for (isize i = 0; i < len; i++) {
-        hash ^= cast(byte)data[i];
+        hash ^= cast(u32)data[i];
         hash *= FNV1A_PRIME_32;
     }
     return hash;
